Better naming and file separation for toolbar objects
