+++ Client/Frontend/Browser/Toolbars/ToolbarButton.swift
@@ -1,0 +1,53 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+import UIKit
+
+class ToolbarButton: UIButton {
+    var selectedTintColor: UIColor!
+    var unselectedTintColor: UIColor!
+    var disabledTintColor: UIColor!
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        adjustsImageWhenHighlighted = false
+        selectedTintColor = tintColor
+        unselectedTintColor = tintColor
+        disabledTintColor = UIColor.Photon.Grey50
+        imageView?.contentMode = .scaleAspectFit
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    override open var isHighlighted: Bool {
+        didSet {
+            self.tintColor = isHighlighted ? selectedTintColor : unselectedTintColor
+        }
+    }
+    
+    override open var isEnabled: Bool {
+        didSet {
+            self.tintColor = isEnabled ? unselectedTintColor : disabledTintColor
+        }
+    }
+    
+    override var tintColor: UIColor! {
+        didSet {
+            self.imageView?.tintColor = self.tintColor
+        }
+    }
+    
+}
+
+extension ToolbarButton: Themeable {
+    func applyTheme(_ theme: Theme) {
+        selectedTintColor = UIColor.ToolbarButton.SelectedTint.colorFor(theme)
+        disabledTintColor = UIColor.ToolbarButton.DisabledTint.colorFor(theme)
+        unselectedTintColor = UIColor.Browser.Tint.colorFor(theme)
+        tintColor = isEnabled ? unselectedTintColor : disabledTintColor
+        imageView?.tintColor = tintColor
+    }
+}
